cogs/math: remove link previews for problem statements and solutions
Currently, a link preview is sent for problem statements, but there are no metadata tags
on the raw image, so a blank preview is shown. As for AOPS solution links, the mediawiki
metadata is not being picked up by discord, hence only the site root is shown.

This commit disables the link preview functionality by enclosing links in `<>`. Some
unnecessary while loops were also removed for clarity.